requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
docopt==0.6.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
loguru==0.7.3
multidict==6.7.1
pipreqs==0.4.13
propcache==0.4.1
python-telegram-bot==22.7
PyYAML==6.0.3
requests==2.32.5
setuptools==80.10.2
urllib3==2.6.3
wheel==0.46.3
yarg==0.1.10
yarl==1.23.0
